Revert "Revert "Remove dataflow_test.py""
This reverts commit 96fcc7d31c2540f867c3a73903c2aa99183a6b8b.
